% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/activate.R
\name{activate}
\alias{activate}
\alias{info}
\alias{is.activated}
\title{Activate a virtual environment}
\usage{
activate(path = NULL)

is.activated()

info()
}
\arguments{
\item{path(string)}{}
}
\description{
Activate a virtual environment
}

