% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_use_local_lib.R
\name{cluster_use_local_lib}
\alias{cluster_use_local_lib}
\title{Activate a local library remotely.}
\usage{
cluster_use_local_lib(.cl)
}
\arguments{
\item{.cl(cluster)}{A \code{multidplyr} cluster.}
}
\value{
\code{cluster}: The cluster \code{.cl} with the locally activated local library
  activated on each of the remote nodes.
}
\description{
This function activates the current local library activated on each of the nodes of
a cluster.
}
\section{Errors}{

\itemize{
  \item{argument \code{.cl} does not inherit from class \code{cluster}}
  \item{locally library not activated locally}
  \item{namespace \code{multidplyr} cannot be loaded}
}
}

