% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-Rprofile.R
\name{create_.Rprofile}
\alias{create_.Rprofile}
\title{Create a .Rprofile}
\usage{
create_.Rprofile(path = NULL)
}
\description{
Create a .Rprofile which creates a local library, and then activates it
(if R is being used interactively).
}

