% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/global-libs.R
\name{create_global_lib}
\alias{create_global_lib}
\title{Create a global library}
\usage{
create_global_lib(name)
}
\arguments{
\item{name}{}
}
\description{
Create a global library
}
\details{
Requires the environment variable `LOCALLIB_HOME` to be set
}

