% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create-local-lib.R
\name{create_local_lib}
\alias{create_local_lib}
\title{Create a local library}
\usage{
create_local_lib(path = NULL)
}
\arguments{
\item{path(string)}{Path of directory to create the local library in.}
}
\description{
Create a local library
}

