% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drat-lib-pkgs.R
\name{drat_insert}
\alias{drat_insert}
\alias{drat_insert_lib}
\alias{drat_insert_pkgs}
\title{drat insert}
\usage{
drat_insert_lib(path, commit = TRUE)

drat_insert_pkgs(path, commit = TRUE)
}
\description{
Insert packages into a local drat repo from either a library or a pkgs.yaml file.
}

