% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drat-install.R
\name{drat_install}
\alias{drat_install}
\title{drat install}
\usage{
drat_install(...)
}
\description{
Install a package into the virtual environment, then insert the source for the
installed packge into the drat repository.
}

