% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/install-version.R
\name{install_version}
\alias{install_version}
\title{Install a particular version of a package}
\usage{
install_version(pkg, version)
}
\arguments{
\item{pkg(string)}{Package name.}

\item{version(string)}{Package version.}
}
\description{
Install a particular version of a package
}

