% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is-installed-locally.R
\name{is.installed_locally}
\alias{is.installed_locally}
\title{Is a package installed locally}
\usage{
is.installed_locally(pkg, version = NULL)
}
\arguments{
\item{pkg(string)}{Name of package.}

\item{version(string)}{Package version. If \code{NULL}, function checks to see if any
version of the package is installed.}
}
\description{
Utility function to see if a package is installed locally.
}

