% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_util.R
\name{is_string_with_content}
\alias{is_string_with_content}
\title{check that a value is a plain string that is not empty}
\usage{
is_string_with_content(input)
}
\arguments{
\item{input}{An arbitrary value to check}
}
\description{
check that a value is a plain string that is not empty
}
