% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recreate-local-lib.R
\name{recreate_local_lib}
\alias{recreate_local_lib}
\title{Recreate a local library}
\usage{
recreate_local_lib(path = NULL)
}
\arguments{
\item{path(string)}{Path of directory to create the local library in.}
}
\description{
Recreate a local library
}

