% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/thaw.R
\name{thaw}
\alias{thaw}
\title{Recreate a virtual environment}
\usage{
thaw(...)
}
\arguments{
\item{...}{Further arguments to pass to \code{\link{install.packages}()}
 (not the package name).}
}
\description{
Recreate a virtual environment
}

