% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/global-libs.R
\name{use_global_lib}
\alias{use_global_lib}
\title{Activate global library}
\usage{
use_global_lib(name)
}
\arguments{
\item{name(string)}{Name of the global library}
}
\description{
Activate global library
}

