% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/use-local-lib.R
\name{use_local_lib}
\alias{use_local_lib}
\title{Activate a local library}
\usage{
use_local_lib(path = NULL)
}
\arguments{
\item{path(string)}{Path to directory containing a local library.}
}
\description{
Activate a local library
}

