% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/virtualenv.R
\name{virtualenv}
\alias{virtualenv}
\title{Create a virtual environment}
\usage{
virtualenv(path = NULL)
}
\arguments{
\item{path(string)}{File path of the virtual environment.}
}
\description{
Create a virtual environment
}

