% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/activate.R
\name{activate}
\alias{activate}
\alias{is.activated}
\title{Activate a virtual environment}
\usage{
activate(path = NULL, deps = TRUE, drat = TRUE, ...)

is.activated()
}
\arguments{
\item{path(string)}{}
}
\description{
Activate a virtual environment
}

