% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drat-install-update.R
\name{drat_install}
\alias{drat_install}
\alias{drat_update}
\title{drat install}
\usage{
drat_install(..., commit = FALSE)

drat_update(..., commit = FALSE)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{install.packages}()} or
\code{\link{update.packages}()} respectively.}

\item{commit}{See \code{commit} argument of \code{drat::\link[drat]{insertPackage}()}.}
}
\description{
Install packages into the virtual environment, then insert the source for the
installed packge into the drat repository.
}

