% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-version.R
\name{install_version}
\alias{install_version}
\title{Install a particular version of a package into a local library}
\usage{
install_version(pkg, version = NULL)
}
\arguments{
\item{pkg(string)}{Package name.}

\item{version(string)}{Package version.}
}
\description{
Install a particular version of a package into a local library
}

